+++ src/main.py
@@ -1,6 +1,6 @@
 import gui
+import database
 
 
-
-
+database.create_database()
 gui.start()